requirements.txt
causal_conv1d==1.4.0
einops==0.8.2
ftfy==6.3.1
h5py==3.15.1
lm_eval==0.4.9.2
mamba_ssm==2.2.4
numpy==2.4.1
packaging==26.0
Pillow==10.0.1
Pillow==12.1.0
pytest==9.0.1
regex==2023.12.25
scipy==1.17.0
setuptools==80.7.1
smplx==0.1.28
timm==1.0.24
torch==2.3.1+cu118
torchvision==0.18.1+cu118
tqdm==4.66.6
transformers==4.38.2
wheel==0.46.3
yacs==0.1.8
